backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "garmintracker-backend"
version = "1.0.0"
description = "GarminTracker backend — health platform connecting patients with doctors through Garmin wearable data and AI"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "pydantic-settings>=2.7",
    "python-jose[cryptography]>=3.3",
    "cryptography>=44.0",
    "garminconnect>=0.2.25",
    "anthropic>=0.43",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "python-multipart>=0.0.18",
    "slowapi>=0.1.9",
    "boto3>=1.35",
    "stripe>=11.0",
    "sse-starlette>=2.0",
    "structlog>=24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.8",
    "mypy>=1.13",
    "aiosqlite>=0.20",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]
